end section 13

+++ src/app/api/todos/route.ts
@@ -1,6 +1,7 @@
 import prisma from "@/lib/prisma";
 import { NextResponse} from "next/server";
-import * as yup from "yup";
+import * as yup from "yup"; // authentication package
+import { auth } from "../../../../auth";
 
 export async function GET(request: Request) {
   const { searchParams } = new URL(request.url);
@@ -35,10 +36,17 @@ const postSchema = yup.object({
 
 export async function POST(request: Request) { 
 
+  const session = await auth();
+  const user = session?.user
+
+  if (!user) {
+    return NextResponse.json({ message: "Unauthorized" }, { status: 401 });
+  }
+
   try {
     const  {complete, description}  = await postSchema.validate( await request.json() );
 
-    const todo = await prisma.todo.create({ data:  {complete, description}  })
+    const todo = await prisma.todo.create({ data:  {complete, description, userId: user.id}  })
   
     
     return NextResponse.json(todo);
@@ -51,10 +59,17 @@ export async function POST(request: Request) {
 
 export async function DELETE() {
 
+  const session = await auth();
+  const user = session?.user
+
+  if (!user) {
+    return NextResponse.json({ message: "Unauthorized" }, { status: 401 });
+  }
+
   try {
     
     await prisma.todo.deleteMany({
-      where: { complete: true},
+      where: { complete: true, userId: user.id},
     });
 
     return NextResponse.json({ message: "Deleted completed todos" });
